+++ src/main/java/lesson5/task2/Triangle.java
@@ -1,14 +1,14 @@
 package lesson5.task2;
 
 class Triangle implements Shape {
-    private int sideA;
-    private int sideB;
-    private int sideC;
+    private double sideA;
+    private double sideB;
+    private double sideC;
 
     private String fillColor;
     private String borderColor;
 
-    public Triangle(int sideA, int sideB, int sideC, String fillColor, String borderColor) {
+    public Triangle(double sideA, double sideB, double sideC, String fillColor, String borderColor) {
         this.sideA = sideA;
         this.sideB = sideB;
         this.sideC = sideC;
